Change shininess to integer
- Fixed lighting issue
- Should also help with performance

+++ src/materials.rs
@@ -3,7 +3,7 @@ use crate::lighting::Color;
 pub struct Material {
     pub diffusion: f32,
     pub specularity: f32,
-    pub shininess: f32,
+    pub shininess: i32,
     pub reflectance: f32,
     pub color: Color,
 }
